Many more instances of ints being compared to size_ts fixed

+++ include/command.tcc
@@ -32,7 +32,7 @@
 template <class T>
 void Command::add_fields(const std::vector<T>& fields, bool is_key)
 {
-    for(int i=0; i<fields.size(); i++) {
+    for (size_t i = 0; i < fields.size(); i++) {
         this->add_field(std::to_string(fields[i]), is_key);
     }
     return;
